Create code 912
This application is used to predict the weather of the present day and also helps the user to get an overview of the weather conditions for a few upcoming days and schedule their important work in accordance with it. can use any weather API (Application Programming Interface) to fetch information about the weather of any city.ncan also fetch other environmental details like Humidity, AQI, temperature, etc. The prerequisites for this project are Django and Python. One of the best APIs is Open Weather API